update package refs
