+++ backend/config_test.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend/proxy"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/useragent"
+	"github.com/grafana/grafana-plugin-sdk-go/config"
 	"github.com/grafana/grafana-plugin-sdk-go/experimental/featuretoggles"
 )
 
@@ -21,37 +22,37 @@ func TestConfig(t *testing.T) {
 	t.Run("GrafanaConfigFromContext", func(t *testing.T) {
 		tcs := []struct {
 			name                   string
-			cfg                    *GrafanaCfg
-			expectedFeatureToggles FeatureToggles
-			expectedProxy          Proxy
+			cfg                    *config.GrafanaCfg
+			expectedFeatureToggles config.FeatureToggles
+			expectedProxy          config.Proxy
 		}{
 			{
 				name:                   "nil config",
 				cfg:                    nil,
-				expectedFeatureToggles: FeatureToggles{},
-				expectedProxy:          Proxy{},
+				expectedFeatureToggles: config.NewGrafanaCfg(nil).FeatureToggles(),
+				expectedProxy:          config.Proxy{},
 			},
 			{
 				name:                   "empty config",
-				cfg:                    &GrafanaCfg{},
-				expectedFeatureToggles: FeatureToggles{},
-				expectedProxy:          Proxy{},
+				cfg:                    &config.GrafanaCfg{},
+				expectedFeatureToggles: config.NewGrafanaCfg(nil).FeatureToggles(),
+				expectedProxy:          config.Proxy{},
 			},
 			{
 				name:                   "nil config map",
-				cfg:                    NewGrafanaCfg(nil),
-				expectedFeatureToggles: FeatureToggles{},
-				expectedProxy:          Proxy{},
+				cfg:                    config.NewGrafanaCfg(nil),
+				expectedFeatureToggles: config.NewGrafanaCfg(nil).FeatureToggles(),
+				expectedProxy:          config.Proxy{},
 			},
 			{
 				name:                   "empty config map",
-				cfg:                    NewGrafanaCfg(make(map[string]string)),
-				expectedFeatureToggles: FeatureToggles{},
-				expectedProxy:          Proxy{},
+				cfg:                    config.NewGrafanaCfg(make(map[string]string)),
+				expectedFeatureToggles: config.NewGrafanaCfg(nil).FeatureToggles(),
+				expectedProxy:          config.Proxy{},
 			},
 			{
 				name: "feature toggles and proxy enabled",
-				cfg: NewGrafanaCfg(map[string]string{
+				cfg: config.NewGrafanaCfg(map[string]string{
 					featuretoggles.EnabledFeatures:           "TestFeature",
 					proxy.PluginSecureSocksProxyEnabled:      "true",
 					proxy.PluginSecureSocksProxyProxyAddress: "localhost:1234",
@@ -60,13 +61,11 @@ func TestConfig(t *testing.T) {
 					proxy.PluginSecureSocksProxyClientCert:   "clientCert",
 					proxy.PluginSecureSocksProxyRootCAs:      "rootCACert",
 				}),
-				expectedFeatureToggles: FeatureToggles{
-					enabled: map[string]struct{}{
-						"TestFeature": {},
-					},
-				},
-				expectedProxy: Proxy{
-					clientCfg: &proxy.ClientCfg{
+				expectedFeatureToggles: config.NewGrafanaCfg(map[string]string{
+					featuretoggles.EnabledFeatures: "TestFeature",
+				}).FeatureToggles(),
+				expectedProxy: config.Proxy{
+					ClientCfg: &proxy.ClientCfg{
 						ClientCert:   "clientCert",
 						ClientKey:    "clientKey",
 						RootCAs:      []string{"rootCACert"},
@@ -77,7 +76,7 @@ func TestConfig(t *testing.T) {
 			},
 			{
 				name: "feature toggles enabled and proxy disabled",
-				cfg: NewGrafanaCfg(map[string]string{
+				cfg: config.NewGrafanaCfg(map[string]string{
 					featuretoggles.EnabledFeatures:           "TestFeature",
 					proxy.PluginSecureSocksProxyEnabled:      "false",
 					proxy.PluginSecureSocksProxyProxyAddress: "localhost:1234",
@@ -86,16 +85,14 @@ func TestConfig(t *testing.T) {
 					proxy.PluginSecureSocksProxyClientCert:   "clientCert",
 					proxy.PluginSecureSocksProxyRootCAs:      "rootCACert",
 				}),
-				expectedFeatureToggles: FeatureToggles{
-					enabled: map[string]struct{}{
-						"TestFeature": {},
-					},
-				},
-				expectedProxy: Proxy{},
+				expectedFeatureToggles: config.NewGrafanaCfg(map[string]string{
+					featuretoggles.EnabledFeatures: "TestFeature",
+				}).FeatureToggles(),
+				expectedProxy: config.Proxy{},
 			},
 			{
 				name: "feature toggles disabled and proxy enabled",
-				cfg: NewGrafanaCfg(map[string]string{
+				cfg: config.NewGrafanaCfg(map[string]string{
 					featuretoggles.EnabledFeatures:           "",
 					proxy.PluginSecureSocksProxyEnabled:      "true",
 					proxy.PluginSecureSocksProxyProxyAddress: "localhost:1234",
@@ -104,9 +101,11 @@ func TestConfig(t *testing.T) {
 					proxy.PluginSecureSocksProxyClientCert:   "clientCert",
 					proxy.PluginSecureSocksProxyRootCAs:      "rootCACert",
 				}),
-				expectedFeatureToggles: FeatureToggles{},
-				expectedProxy: Proxy{
-					clientCfg: &proxy.ClientCfg{
+				expectedFeatureToggles: config.NewGrafanaCfg(map[string]string{
+					featuretoggles.EnabledFeatures: "",
+				}).FeatureToggles(),
+				expectedProxy: config.Proxy{
+					ClientCfg: &proxy.ClientCfg{
 						ClientCert:   "clientCert",
 						ClientKey:    "clientKey",
 						RootCAs:      []string{"rootCACert"},
@@ -117,7 +116,7 @@ func TestConfig(t *testing.T) {
 			},
 			{
 				name: "feature toggles disabled and insecure proxy enabled",
-				cfg: NewGrafanaCfg(map[string]string{
+				cfg: config.NewGrafanaCfg(map[string]string{
 					featuretoggles.EnabledFeatures:            "",
 					proxy.PluginSecureSocksProxyEnabled:       "true",
 					proxy.PluginSecureSocksProxyProxyAddress:  "localhost:1234",
@@ -127,9 +126,11 @@ func TestConfig(t *testing.T) {
 					proxy.PluginSecureSocksProxyRootCAs:       "rootCACert",
 					proxy.PluginSecureSocksProxyAllowInsecure: "true",
 				}),
-				expectedFeatureToggles: FeatureToggles{},
-				expectedProxy: Proxy{
-					clientCfg: &proxy.ClientCfg{
+				expectedFeatureToggles: config.NewGrafanaCfg(map[string]string{
+					featuretoggles.EnabledFeatures: "",
+				}).FeatureToggles(),
+				expectedProxy: config.Proxy{
+					ClientCfg: &proxy.ClientCfg{
 						ClientCert:    "clientCert",
 						ClientKey:     "clientKey",
 						RootCAs:       []string{"rootCACert"},
@@ -141,7 +142,7 @@ func TestConfig(t *testing.T) {
 			},
 			{
 				name: "feature toggles disabled and secure proxy enabled with file contents",
-				cfg: NewGrafanaCfg(map[string]string{
+				cfg: config.NewGrafanaCfg(map[string]string{
 					featuretoggles.EnabledFeatures:                 "",
 					proxy.PluginSecureSocksProxyEnabled:            "true",
 					proxy.PluginSecureSocksProxyProxyAddress:       "localhost:1234",
@@ -154,9 +155,11 @@ func TestConfig(t *testing.T) {
 					proxy.PluginSecureSocksProxyRootCAsContents:    "rootCACert,rootCACert2",
 					proxy.PluginSecureSocksProxyAllowInsecure:      "true",
 				}),
-				expectedFeatureToggles: FeatureToggles{},
-				expectedProxy: Proxy{
-					clientCfg: &proxy.ClientCfg{
+				expectedFeatureToggles: config.NewGrafanaCfg(map[string]string{
+					featuretoggles.EnabledFeatures: "",
+				}).FeatureToggles(),
+				expectedProxy: config.Proxy{
+					ClientCfg: &proxy.ClientCfg{
 						ClientCert:    "./clientCert",
 						ClientCertVal: "clientCert",
 						ClientKey:     "./clientKey",
@@ -172,11 +175,11 @@ func TestConfig(t *testing.T) {
 		}
 
 		for _, tc := range tcs {
-			ctx := WithGrafanaConfig(context.Background(), tc.cfg)
-			cfg := GrafanaConfigFromContext(ctx)
+			ctx := config.WithGrafanaConfig(context.Background(), tc.cfg)
+			cfg := config.GrafanaConfigFromContext(ctx)
 
 			require.Equal(t, tc.expectedFeatureToggles, cfg.FeatureToggles())
-			proxy, err := cfg.proxy()
+			proxy, err := cfg.Proxy()
 			assert.NoError(t, err)
 
 			require.Equal(t, tc.expectedProxy, proxy)
@@ -186,8 +189,8 @@ func TestConfig(t *testing.T) {
 
 func TestAppURL(t *testing.T) {
 	t.Run("it should return the configured app URL", func(t *testing.T) {
-		cfg := NewGrafanaCfg(map[string]string{
-			AppURL: "http://localhost:3000",
+		cfg := config.NewGrafanaCfg(map[string]string{
+			config.AppURL: "http://localhost:3000",
 		})
 		url, err := cfg.AppURL()
 		require.NoError(t, err)
@@ -195,15 +198,15 @@ func TestAppURL(t *testing.T) {
 	})
 
 	t.Run("it should return an error if the app URL is missing", func(t *testing.T) {
-		cfg := NewGrafanaCfg(map[string]string{})
+		cfg := config.NewGrafanaCfg(map[string]string{})
 		_, err := cfg.AppURL()
 		require.Error(t, err)
 	})
 
 	t.Run("it should return the configured app URL from env", func(t *testing.T) {
-		os.Setenv(AppURL, "http://localhost-env:3000")
-		defer os.Unsetenv(AppURL)
-		cfg := NewGrafanaCfg(map[string]string{})
+		os.Setenv(config.AppURL, "http://localhost-env:3000")
+		defer os.Unsetenv(config.AppURL)
+		cfg := config.NewGrafanaCfg(map[string]string{})
 		v, err := cfg.AppURL()
 		require.NoError(t, err)
 		require.Equal(t, "http://localhost-env:3000", v)
@@ -231,8 +234,8 @@ func TestUserAgentFromContext_NoUserAgent(t *testing.T) {
 
 func TestUserFacingDefaultError(t *testing.T) {
 	t.Run("it should return the configured default error message", func(t *testing.T) {
-		cfg := NewGrafanaCfg(map[string]string{
-			UserFacingDefaultError: "something failed",
+		cfg := config.NewGrafanaCfg(map[string]string{
+			config.UserFacingDefaultError: "something failed",
 		})
 		v, err := cfg.UserFacingDefaultError()
 		require.NoError(t, err)
@@ -240,7 +243,7 @@ func TestUserFacingDefaultError(t *testing.T) {
 	})
 
 	t.Run("it should return an error if the default error message is missing", func(t *testing.T) {
-		cfg := NewGrafanaCfg(map[string]string{})
+		cfg := config.NewGrafanaCfg(map[string]string{})
 		_, err := cfg.UserFacingDefaultError()
 		require.Error(t, err)
 	})
@@ -248,11 +251,11 @@ func TestUserFacingDefaultError(t *testing.T) {
 
 func TestSql(t *testing.T) {
 	t.Run("it should return the configured sql default values", func(t *testing.T) {
-		cfg := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:                      "5",
-			SQLMaxOpenConnsDefault:           "11",
-			SQLMaxIdleConnsDefault:           "22",
-			SQLMaxConnLifetimeSecondsDefault: "33",
+		cfg := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:                      "5",
+			config.SQLMaxOpenConnsDefault:           "11",
+			config.SQLMaxIdleConnsDefault:           "22",
+			config.SQLMaxConnLifetimeSecondsDefault: "33",
 		})
 		v, err := cfg.SQL()
 		require.NoError(t, err)
@@ -263,28 +266,28 @@ func TestSql(t *testing.T) {
 	})
 
 	t.Run("it should return an error if any of the defaults is missing", func(t *testing.T) {
-		cfg1 := NewGrafanaCfg(map[string]string{
-			SQLMaxOpenConnsDefault:           "11",
-			SQLMaxIdleConnsDefault:           "22",
-			SQLMaxConnLifetimeSecondsDefault: "33",
+		cfg1 := config.NewGrafanaCfg(map[string]string{
+			config.SQLMaxOpenConnsDefault:           "11",
+			config.SQLMaxIdleConnsDefault:           "22",
+			config.SQLMaxConnLifetimeSecondsDefault: "33",
 		})
 
-		cfg2 := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:                      "5",
-			SQLMaxIdleConnsDefault:           "22",
-			SQLMaxConnLifetimeSecondsDefault: "33",
+		cfg2 := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:                      "5",
+			config.SQLMaxIdleConnsDefault:           "22",
+			config.SQLMaxConnLifetimeSecondsDefault: "33",
 		})
 
-		cfg3 := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:                      "5",
-			SQLMaxOpenConnsDefault:           "11",
-			SQLMaxConnLifetimeSecondsDefault: "33",
+		cfg3 := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:                      "5",
+			config.SQLMaxOpenConnsDefault:           "11",
+			config.SQLMaxConnLifetimeSecondsDefault: "33",
 		})
 
-		cfg4 := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:            "5",
-			SQLMaxOpenConnsDefault: "11",
-			SQLMaxIdleConnsDefault: "22",
+		cfg4 := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:            "5",
+			config.SQLMaxOpenConnsDefault: "11",
+			config.SQLMaxIdleConnsDefault: "22",
 		})
 
 		_, err := cfg1.SQL()
@@ -301,32 +304,32 @@ func TestSql(t *testing.T) {
 	})
 
 	t.Run("it should return an error if any of the defaults is not an integer", func(t *testing.T) {
-		cfg1 := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:                      "not-an-integer",
-			SQLMaxOpenConnsDefault:           "11",
-			SQLMaxIdleConnsDefault:           "22",
-			SQLMaxConnLifetimeSecondsDefault: "33",
+		cfg1 := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:                      "not-an-integer",
+			config.SQLMaxOpenConnsDefault:           "11",
+			config.SQLMaxIdleConnsDefault:           "22",
+			config.SQLMaxConnLifetimeSecondsDefault: "33",
 		})
 
-		cfg2 := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:                      "5",
-			SQLMaxOpenConnsDefault:           "11",
-			SQLMaxIdleConnsDefault:           "not-an-integer",
-			SQLMaxConnLifetimeSecondsDefault: "33",
+		cfg2 := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:                      "5",
+			config.SQLMaxOpenConnsDefault:           "11",
+			config.SQLMaxIdleConnsDefault:           "not-an-integer",
+			config.SQLMaxConnLifetimeSecondsDefault: "33",
 		})
 
-		cfg3 := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:                      "5",
-			SQLMaxOpenConnsDefault:           "11",
-			SQLMaxIdleConnsDefault:           "not-an-integer",
-			SQLMaxConnLifetimeSecondsDefault: "33",
+		cfg3 := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:                      "5",
+			config.SQLMaxOpenConnsDefault:           "11",
+			config.SQLMaxIdleConnsDefault:           "not-an-integer",
+			config.SQLMaxConnLifetimeSecondsDefault: "33",
 		})
 
-		cfg4 := NewGrafanaCfg(map[string]string{
-			SQLRowLimit:                      "5",
-			SQLMaxOpenConnsDefault:           "11",
-			SQLMaxIdleConnsDefault:           "22",
-			SQLMaxConnLifetimeSecondsDefault: "not-an-integer",
+		cfg4 := config.NewGrafanaCfg(map[string]string{
+			config.SQLRowLimit:                      "5",
+			config.SQLMaxOpenConnsDefault:           "11",
+			config.SQLMaxIdleConnsDefault:           "22",
+			config.SQLMaxConnLifetimeSecondsDefault: "not-an-integer",
 		})
 
 		_, err := cfg1.SQL()
@@ -345,8 +348,8 @@ func TestSql(t *testing.T) {
 
 func TestPluginAppClientSecret(t *testing.T) {
 	t.Run("it should return the configured PluginAppClientSecret", func(t *testing.T) {
-		cfg := NewGrafanaCfg(map[string]string{
-			AppClientSecret: "client-secret",
+		cfg := config.NewGrafanaCfg(map[string]string{
+			config.AppClientSecret: "client-secret",
 		})
 		v, err := cfg.PluginAppClientSecret()
 		require.NoError(t, err)
@@ -354,9 +357,9 @@ func TestPluginAppClientSecret(t *testing.T) {
 	})
 
 	t.Run("it should return the configured PluginAppClientSecret from env", func(t *testing.T) {
-		os.Setenv(AppClientSecret, "client-secret")
-		defer os.Unsetenv(AppClientSecret)
-		cfg := NewGrafanaCfg(map[string]string{})
+		os.Setenv(config.AppClientSecret, "client-secret")
+		defer os.Unsetenv(config.AppClientSecret)
+		cfg := config.NewGrafanaCfg(map[string]string{})
 		v, err := cfg.PluginAppClientSecret()
 		require.NoError(t, err)
 		require.Equal(t, "client-secret", v)
@@ -384,7 +387,7 @@ func BenchmarkProxyHash(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		kBytes[88] = b64chars[mathrand.Intn(64)] //nolint:gosec
 		cm[proxy.PluginSecureSocksProxyClientKeyContents] = string(kBytes)
-		cfg := NewGrafanaCfg(cm)
+		cfg := config.NewGrafanaCfg(cm)
 		hash := cfg.ProxyHash()
 		if hash[0] == 'a' {
 			count++
